demos/04-hardware-security/requirements.txt
avp-sdk>=0.1.0
pyserial>=3.5
anthropic>=0.18.0
